LINUX - fixed timing cast

+++ src/ccore/linux/utils/lin_time.c
@@ -5,9 +5,9 @@ void ccTimeDelay(int ms)
 	usleep(ms * _TO_MICROSECONDS);
 }
 
-unsigned long ccTimeNanoseconds()
+uint64_t ccTimeNanoseconds()
 {
 	struct timespec time;
 	clock_gettime(CLOCK_REALTIME, &time);
-	return time.tv_nsec + time.tv_sec * _TO_SECONDS;
+	return (uint64_t)(time.tv_nsec + time.tv_sec * _TO_SECONDS);
 }
